+++ CRM/Dataprocessor/Utils/DataSourceFields.php
@@ -56,7 +56,7 @@ class CRM_Dataprocessor_Utils_DataSourceFields {
    * Returns an array with the name of the field as the key and the label of the field as the value.
    *
    * @oaram int $dataProcessorId
-   * @param callable $callback
+   * @param callable $filterFieldsCallback
    *   Function to filter certain fields.
    * @return array
    * @throws \Exception
@@ -66,7 +66,7 @@ class CRM_Dataprocessor_Utils_DataSourceFields {
     $dataProcessorClass = \CRM_Dataprocessor_BAO_DataProcessor::dataProcessorToClass($dataProcessor);
     $fieldSelect = array();
     foreach($dataProcessorClass->getDataSources() as $dataSource) {
-      $fieldSelect = array_merge($fieldSelect, self::getAvailableFilterFieldsInDataSource($dataSource, $dataSource->getSourceTitle().' :: ', $dataSource->getSourceName().'::', $filterFieldsCallback));
+      $fieldSelect = array_merge($fieldSelect, self::getAvailableFilterFieldsInDataSource($dataSource, $dataSource->getSourceTitle() . ' :: ', $dataSource->getSourceName() . '::', $filterFieldsCallback));
     }
     return $fieldSelect;
   }
@@ -87,7 +87,7 @@ class CRM_Dataprocessor_Utils_DataSourceFields {
     foreach($dataSource->getAvailableFilterFields()->getFields() as $field) {
       $isFieldValid = true;
       if ($filterFieldsCallback) {
-        $isFieldValid = call_user_func($filterFieldsCallback, $field);
+        $isFieldValid = call_user_func($filterFieldsCallback, $field, $dataSource);
       }
       if ($isFieldValid) {
         $fieldSelect[$namePrefix . $field->alias] = $titlePrefix . $field->title;