+++ Civi/DataProcessor/DataFlow/InMemoryDataFlow.php
@@ -6,6 +6,8 @@
 
 namespace Civi\DataProcessor\DataFlow;
 
+use Civi\DataProcessor\DataFlow\InMemoryDataFlow\FilterInterface;
+
 class InMemoryDataFlow extends AbstractDataFlow {
 
   protected $data = [];
@@ -14,7 +16,12 @@ class InMemoryDataFlow extends AbstractDataFlow {
 
   protected $isInitialized = FALSE;
 
-  public function __construct($data) {
+  /**
+   * @var \Civi\DataProcessor\DataFlow\InMemoryDataFlow\FilterInterface[]
+   */
+  protected $filters = [];
+
+  public function __construct($data=null) {
     parent::__construct();
     $this->data = $data;
   }
@@ -42,7 +49,7 @@ class InMemoryDataFlow extends AbstractDataFlow {
     return $this->isInitialized;
   }
 
-/**
+  /**
    * Resets the initialized state. This function is called
    * when a setting has changed. E.g. when offset or limit are set.
    *
@@ -62,18 +69,38 @@ class InMemoryDataFlow extends AbstractDataFlow {
    * @throws EndOfFlowException
    */
   public function retrieveNextRecord($fieldNameprefix='') {
-    if (!isset($this->data[$this->currentPointer])) {
-      throw new EndOfFlowException();
+    if (!$this->isInitialized()) {
+      $this->initialize();
     }
-    $data = $this->data[$this->currentPointer];
-    $record = array();
-    foreach($this->dataSpecification->getFields() as $field) {
-      $alias = $field->alias;
-      $name = $field->name;
-      $record[$fieldNameprefix.$alias] = $data[$name];
-    }
-    $this->currentPointer++;
+    do {
+      if (!isset($this->data[$this->currentPointer])) {
+        throw new EndOfFlowException();
+      }
+      $data = $this->data[$this->currentPointer];
+      $record = [];
+      foreach ($this->dataSpecification->getFields() as $field) {
+        $alias = $field->alias;
+        $name = $field->name;
+        $record[$fieldNameprefix . $alias] = $data[$name];
+      }
+      $this->currentPointer++;
+    } while (!$this->filterRecord($record));
+
     return $record;
+  }
+
+  /**
+   * @param $record
+   *
+   * @return bool
+   */
+  protected function filterRecord($record) {
+    foreach($this->filters as $filter) {
+      if (!$filter->filterRecord($record)) {
+        return FALSE;
+      }
+    }
+    return TRUE;
   }
 
   /**
@@ -83,6 +110,36 @@ class InMemoryDataFlow extends AbstractDataFlow {
    */
   public function getName() {
     return 'in_memory';
+  }
+
+  /**
+   * Adds a filter to the data flow.
+   *
+   * @param \Civi\DataProcessor\DataFlow\InMemoryDataFlow\FilterInterface $filter
+   */
+  public function addFilter(FilterInterface $filter) {
+    if (!in_array($filter, $this->filters)) {
+      $this->filters[] = $filter;
+    }
+  }
+
+  /**
+   * Removes a filter from the data flow.
+   *
+   * @param \Civi\DataProcessor\DataFlow\InMemoryDataFlow\FilterInterface $filter
+   */
+  public function removeFilter(FilterInterface $filter) {
+    $key = array_search($filter, $this->filters);
+    if ($key!==false) {
+      unset($this->filters[$key]);
+    }
+  }
+
+  /**
+   * @return \Civi\DataProcessor\DataFlow\InMemoryDataFlow\FilterInterface[]
+   */
+  public function getFilters() {
+    return $this->filters;
   }
 
 }