+++ Civi/DataProcessor/FilterHandler/AbstractFieldFilterHandler.php
@@ -6,6 +6,7 @@
 
 namespace Civi\DataProcessor\FilterHandler;
 
+use Civi\DataProcessor\DataFlow\InMemoryDataFlow;
 use Civi\DataProcessor\DataFlow\SqlDataFlow;
 use Civi\DataProcessor\Exception\DataSourceNotFoundException;
 use Civi\DataProcessor\Exception\FieldNotFoundException;
@@ -32,6 +33,12 @@ abstract class AbstractFieldFilterHandler extends AbstractFilterHandler {
    * @var \Civi\DataProcessor\DataFlow\SqlDataFlow\WhereClauseInterface
    */
   protected $whereClause;
+
+
+  /**
+   * @var \Civi\DataProcessor\DataFlow\InMemoryDataFlow\SimpleFilter
+   */
+  protected $filterClass;
 
   /**
    * @param $datasource_name
@@ -83,6 +90,9 @@ abstract class AbstractFieldFilterHandler extends AbstractFilterHandler {
     if ($dataFlow && $dataFlow instanceof SqlDataFlow && $this->whereClause) {
       $dataFlow->removeWhereClause($this->whereClause);
       unset($this->whereClause);
+    } elseif ($dataFlow && $dataFlow instanceof InMemoryDataFlow && $this->filterClass) {
+      $dataFlow->removeFilter($this->filterClass);
+      unset($this->filter);
     }
   }
 
@@ -102,6 +112,9 @@ abstract class AbstractFieldFilterHandler extends AbstractFilterHandler {
       }
       $this->whereClause = new SqlDataFlow\SimpleWhereClause($dataFlow->getName(), $this->inputFieldSpecification->name, $filter['op'], $value, $this->inputFieldSpecification->type);
       $dataFlow->addWhereClause($this->whereClause);
+    } elseif ($dataFlow && $dataFlow instanceof InMemoryDataFlow) {
+      $this->filterClass = new InMemoryDataFlow\SimpleFilter($this->inputFieldSpecification->name, $filter['op'], $filter['value']);
+      $dataFlow->addFilter($this->filterClass);
     }
   }